File cleanup and reorganisation: - removed 'data-documents' directory, which contained very old prototypes for component travellers and off-DB JSON files, as well as user-specific files that don't belong in the DB code - removed directory containing 'simple' pug files, which are no longer needed since it is now much easier to prototype and test changes to the main pug files directly ... removed corresponding function in the main app file - renamed the 'static/pages.js' directory to 'static/pages' ... having a directory name with an extension was really bugging me!  Changed references to the directory in pug files as required - moved the 'scss' directory into the 'static' directory, so that it is a 'sibling' to the existing 'css' one ... changed scss compiler section of the main app file accordingly - cleaned up code in 'static/pages/edit_form.js' (included in this commit since that file was also renamed as above) - cleaned up code in 'pug/dashboard.pug' (included in this commit since that file was already modified as above)
