+++ app/pug/EditCourse.pug
@@ -12,7 +12,7 @@ block extrascripts
     var course=!{JSON.stringify((course||{}), null ,4)}; 
     var readOnly=!{retrieved||false};
 
-  script(type='text/javascript',src='/page.js/EditCourse.js')
+  script(type='text/javascript',src='/pages/EditCourse.js')
     
 
 block content
