+++ app/pug/docEdit.pug
@@ -29,7 +29,7 @@ block extrascripts
       }
 
 
-    script(src='/page.js/docEdit.js')
+    script(src='/pages/docEdit.js')
  
 block content
     div.container#wrapper
