+++ app/pug/edit_form.pug
@@ -21,7 +21,7 @@ block extrascripts
   block metaschema
   
   block workscripts
-    script(src = '/page.js/edit_form.js')
+    script(src = '/pages/edit_form.js')
 
 
 block content
