--- app/pug/simple/cy.pug
@@ -1,13 +1,0 @@
-html
-  head
-    script(src='/dist/cytoscape/cytoscape.min.js')
-    script(src='/dist/jquery/jquery.min.js')
-    script(src='/js/cy.js')
-
-    style.
-      #cy {
-       width: 100%;
-       height: 500px;  
-      }
-  body
-    #cy
