--- app/pug/simple/d3-1.pug
@@ -1,44 +1,0 @@
-html
-  head
-  script(src='/dist/jquery/jquery.min.js')
-  script(src='/dist/d3/d3.min.js')
-  script(src='/dist/d3-color-legend/d3-color-legend.min.js')
-  script(src='/js/d3-1.js')
-
-
-    style.
-      #svg-container {
-        display: inline-block;
-        position: relative;
-        width: 100%;
-        padding-bottom:100%;
-        vertical-align: top;
-        overflow:hidden;
-      }
-      .svg-content {
-        display: inline-block;
-        position: absolute:
-        top: 0;
-        left: 0;
-      }
-      #svgd3 {
-        width: 100%;
-        height: 100vh;
-      }
-
-      .bar {
-        fill: green;
-      }
-
-      .highlight {
-        fill: red;
-      }
-
-      .title {
-        fill: blue;
-        font-weight: bold;
-      }
-
-  body
-    #svg-container
-      //- svg#svgd3
