--- app/pug/simple/d3.pug
@@ -1,28 +1,0 @@
-html
-  head
-  script(src='/dist/jquery/jquery.min.js')
-  script(src='/dist/d3/d3.min.js')
-  script(src='/dist/d3-color-legend/d3-color-legend.min.js')
-  script(src='/js/d3.js')
-
-
-  style.
-      #svg-container {
-        display: inline-block;
-        position: relative;
-        width: 100%;
-        //- padding-bottom:100%;
-        //- vertical-align: top;
-        overflow:hidden;
-        height: 80vh;
-        border: 2px solid green;
-      }
-      .svg-content {
-        width: 100%;
-        height: 100%;
-      }
-
-
-  body
-    #svg-container
-      //- svg#svgd3
