--- app/pug/simple/dashboard.pug
@@ -1,256 +1,0 @@
-doctype html
-head
-
-    block pugscripts
-      
-    block vars
-      
-    title #{page_title || 'DUNE DB'}
-
-    <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    //- link(rel='stylesheet' href='/bootstrap/css/bootstrap.min.css')
-    link(rel='icon' type='image/png' href='/images/Otterbein.png')
-
-    link(rel='stylesheet' href='/css/dashboard.css')
-    link(rel='stylesheet' href='/css/custom.scss')
-    link(rel='stylesheet' href='/dist/formiojs/formio.full.min.css')
-
-nav.navbar.navbar-dark.fixed-top.bg-dark.flex-md-nowrap.p-0.shadow
-  a.navbar-brand.col-sm-3.col-md-2.mr-0(href='/') DUNE DB 
-  input.form-control.form-control-dark(type='text' placeholder='Search' aria-label='Search')
-  if locals.user
-    .navbar-nav.px-3
-      .nav-item.text-nowrap
-        a.nav-link(href="/user") Profile
-    .navbar-nav.px-3
-      .nav-item.text-nowrap
-        a.nav-link(href="/logout") Log Out
-  else
-    .navbar-nav.px-3
-        .nav-item.text-nowrap
-          a.nav-link(href="/login") Log In / Sign Up
-
-.container-fluid
-  .row
-    nav.col-md-2.d-none.d-md-block.bg-light.sidebar
-      .sidebar-sticky
-        ul.nav.flex-column
-          li.nav-item
-            .card
-              .card-header
-                .fa.fa-puzzle-piece("aria-hidden"="true")
-                | Components
-              
-
-            a.nav-link(href='#')
-              .fa.fa-puzzle-piece("aria-hidden"="true")
-              | Tests
-
-            a.nav-link(href='#')
-              .fa.fa-puzzle-piece("aria-hidden"="true")
-              | Workflows
-
-            a.nav-link(href='#')
-              | My Drafts
-
-            a.nav-link(href='#')
-              | My recent submissions
-              | mine/others
-
-
-        h6.sidebar-heading.d-flex.justify-content-between.align-items-center.px-3.mt-4.mb-1.text-muted
-          span Admin
-          //- a.d-flex.align-items-center.text-muted(href='#')
-          //-   span(data-feather='plus-circle')
-        ul.nav.flex-column.mb-2
-          li.nav-item
-            a.nav-link(href='#')
-              span(data-feather='file-text')
-              |                   Edit Component Forms
-          li.nav-item
-            a.nav-link(href='#')
-              span(data-feather='file-text')
-              |                   Edit Test Forms
-          li.nav-item
-            a.nav-link(href='#')
-              span(data-feather='file-text')
-              |                   Edit Workflow Forms
-
-
-    main.col-md-9.ml-sm-auto.col-lg-10.px-4(role='main')
-      .d-flex.justify-content-between.flex-wrap.flex-md-nowrap.align-items-center.pt-3.pb-2.mb-3.border-bottom
-        h1.h2 Dashboard
-        .btn-toolbar.mb-2.mb-md-0
-          .btn-group.mr-2
-            button.btn.btn-sm.btn-outline-secondary Share
-            button.btn.btn-sm.btn-outline-secondary Export
-          button.btn.btn-sm.btn-outline-secondary.dropdown-toggle
-            span(data-feather='calendar')
-            |                 This week
-      canvas#myChart.my-4.w-100(width='900' height='380')
-      h2 Section title
-      .table-responsive
-        table.table.table-striped.table-sm
-          thead
-            tr
-              th #
-              th Header
-              th Header
-              th Header
-              th Header
-          tbody
-            tr
-              td 1,001
-              td Lorem
-              td ipsum
-              td dolor
-              td sit
-            tr
-              td 1,002
-              td amet
-              td consectetur
-              td adipiscing
-              td elit
-            tr
-              td 1,003
-              td Integer
-              td nec
-              td odio
-              td Praesent
-            tr
-              td 1,003
-              td libero
-              td Sed
-              td cursus
-              td ante
-            tr
-              td 1,004
-              td dapibus
-              td diam
-              td Sed
-              td nisi
-            tr
-              td 1,005
-              td Nulla
-              td quis
-              td sem
-              td at
-            tr
-              td 1,006
-              td nibh
-              td elementum
-              td imperdiet
-              td Duis
-            tr
-              td 1,007
-              td sagittis
-              td ipsum
-              td Praesent
-              td mauris
-            tr
-              td 1,008
-              td Fusce
-              td nec
-              td tellus
-              td sed
-            tr
-              td 1,009
-              td augue
-              td semper
-              td porta
-              td Mauris
-            tr
-              td 1,010
-              td massa
-              td Vestibulum
-              td lacinia
-              td arcu
-            tr
-              td 1,011
-              td eget
-              td nulla
-              td Class
-              td aptent
-            tr
-              td 1,012
-              td taciti
-              td sociosqu
-              td ad
-              td litora
-            tr
-              td 1,013
-              td torquent
-              td per
-              td conubia
-              td nostra
-            tr
-              td 1,014
-              td per
-              td inceptos
-              td himenaeos
-              td Curabitur
-            tr
-              td 1,015
-              td sodales
-              td ligula
-              td in
-              td libero
-
-
-//
-  Bootstrap core JavaScript
-  ==================================================
-// Placed at the end of the document so the pages load faster
-
-script(src='/dist/jquery/jquery.min.js')
-script(src='/dist/bootstrap/js/bootstrap.min.js')
-script(src="/dist/bootstrap-autocomplete/latest/bootstrap-autocomplete.min.js")
-script(src='/dist/formiojs/formio.full.js')
-script(src="/dist/short-uuid/short-uuid.min.js")
-script(src='/dist/deepmerge/umd.js')
-script(src='/dist/moment/moment.min.js')
-script(src='/dist/ekko-lightbox/ekko-lightbox.min.js')
-
-//- Used to pop open a small image to full-screen
-script.
-  $(document).on('click', '[data-toggle="lightbox"]', function(event) {event.preventDefault();$(this).ekkoLightbox({alwaysShowClose: true});});
-
-//- QR codes, FIXME replace with an npm package?
-script(src='/dist/jsqr/jsQR.js') 
-script(src="/ext/qrcodegen.js")
-
-script(src='/dist/fabric/fabric.min.js')
-script(src='/dist/fabric-history/index.min.js')
-
-script(type="text/javascript", src="/ext/qrcodegen.js")
-script(type="text/javascript", src="/js/generate_qr_code.js")
-script(src='/components/ComponentGeoTag.js')
-script(src='/components/ComponentUUID.js')
-script(src='/components/ComponentWorkStep.js')
-script(src='/components/ComponentArray.js')
-script(src='/components/ComponentSpecNumber.js')
-script(src='/components/Components_misc.js')
-script(src='/components/DatabaseImage.js')
-script(src='/components/DatabaseFile.js')
-
-script(src='/js/AnnotationCanvas.js')
-script(src='/components/ImageAnnotator.js')
-script(src='/components/AnnotatedImage.js')
-script(src='/components/BetterDataGrid.js')
-
-
-script(src='/js/uuid_autocomplete.js')
-script(src='/js/ABoundObject.js')
-script(src='/js/Pad.js')
-script(src='/js/Histogram.js')
-script(src='/js/HistCanvas.js')
-script(src='/js/ColorScale.js')
-
-link(rel='stylesheet', href='/css/main.css', type='text/css', media='all')
-link(rel='stylesheet', href='/css/AnnotationCanvas.css', type='text/css', media='all')
-style.
-  li.nav-item img { height: 40px;}
-
-block extrascripts
