--- app/pug/simple/fabrictest.pug
@@ -1,143 +1,0 @@
-extend ../default
-
-include ../common.pug
-
-block vars
-  - var page_title=pagename;
-
-block extrascripts
-    style.
-        .AnnotationCanvas {
-            width: 100%;
-            height: 80vh;
-            display: flex;
-            flex-flow: column nowrap;
-            align-items: center;
-            margin:3px;
-        }
-        .AnnotationCanvas canvas {
-            border: solid black 2px;
-        }
-
-        .AnnotationCanvas .control_div {
-            display: flex;
-            flex-flow: row wrap;
-        }
-
-        .dragthing {
-            width: 40px;
-            height: 40px;
-            margin: 4px;
-            border: 1px solid #bbb;
-            display: flex;
-            justify-content: center;
-            cursor: move;
-        }
-        .dragthing img {
-            width:38x;
-            maxheight: 38px;
-        }
-        .palette {
-            width: 100px;
-            display: flex;
-            flex-flow: row wrap;
-        }
-        .palette-box {
-            width: 12px;
-            height: 12px;
-            border: 1px solid #aaa;
-            margin:2px;
-            position: relative;
-            overflow: hidden;
-            cursor: pointer;
-        }
-        .palette-box[data-color='']:after {
-              content:"";
-              position:absolute;
-              border-top:1px solid red;
-              width:40px;
-              transform: rotate(45deg);
-              transform-origin: 0% 0%;
-        }
-
-        .unselectable {
-            -webkit-touch-callout: none; /* iOS Safari */
-            -webkit-user-select: none; /* Safari */
-            -khtml-user-select: none; /* Konqueror HTML */
-            -moz-user-select: none; /* Firefox */
-            -ms-user-select: none; /* Internet Explorer/Edge */
-            user-select: none; /* Non-prefixed version, currently
-                                  supported by Chrome and Opera */
-        }
-        
-    script(src='https://cdnjs.cloudflare.com/ajax/libs/fabric.js/3.6.3/fabric.min.js')
-    script(src='/js/fabric-history.js')
-    script(src='/js/AnnotationCanvas.js')
-    script.
-        $(function(){
-
-            $('.launchAnnotator').on('click',function(){
-                console.log('click',this);
-                var modal = $('#AnnotatorModal')
-                modal.modal("show");
-                modal.on("shown.bs.modal",
-                function(){
-                    console.log("modal should be open now");
-                    var canvas = new AnnotationCanvas(
-                        $('.AnnotationCanvas',modal).get(0),
-                        {img_url:'/file/gridfs/5ecd323531486a50e98cec26'});
-                    // for debugging:
-                    gCanvas = canvas;
-                    $(this).off("shown.bs.modal"); // remove this handler.
-
-                    // handle the save button:
-                    $('#AnnotatorModal .SaveAnnotationCanvas').on('click',
-                    function(){
-                        $(this).off('click'); // remove handler.
-                        // save the image.
-                        var uri = canvas.getImageUri();
-                        var json = canvas.getJSON();
-                        $.post({
-                            url:'/file/gridfsBase64',
-                            data: uri,
-                            contentType:'text/plain',
-                            success: function(result) {
-                                // Ok, at this point we should have everything we Non-prefixed  
-                                console.log('success',json,result);
-                                modal.modal("hide");
-                            }
-                        }).fail(function(err){
-                            console.error(err);
-                            $('#AnnotatorModal .SaveAnnotationCanvas').text("Something went wrong!");
-                        })
-                    })
-
-                });
-             }); 
-        })
-
-
-block content
-    p hi there
-
-    button.btn.btn-primary.launchAnnotator Click Me
-
-    <!-- Modal -->
-    <div class="modal fade" id="AnnotatorModal" data-keyboard="false" data-backdrop="static"  role="dialog" aria-labelledby="AnnotatorModalTitle" aria-hidden="true">
-      <div class="modal-dialog modal-xl modal-dialog-centered" role="document">
-        <div class="modal-content">
-          <div class="modal-body">
-                <div class="AnnotationCanvas"></div>
-          </div>
-          <div class="modal-footer">
-            <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
-            <button type="button" class="btn btn-primary SaveAnnotationCanvas">Save changes</button>
-          </div>
-        </div>
-      </div>
-    </div>
-
-
-    //- .AnnotationCanvas
-
-    //- img#benart.img-fuild(src='/file/gridfs/5ecd323531486a50e98cec26')
