--- data-documents/prototraveller/sietch_win/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "sietch-read-prototraveller",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "dependencies": {
-    "adonis-adodb": "^1.0.2",
-    "database-js": "^3.0.10",
-    "database-js-adodb": "^1.1.3",
-    "node-adodb": "^5.0.3"
-  },
-  "devDependencies": {},
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC"
-}
