--- data-documents/prototraveller/sietch_win/tbl030.json
@@ -1,102 +1,0 @@
-[
-  {
-    "APAID": null,
-    "ID": 4,
-    "CR_Boards_Cryo": null,
-    "Wiring_Harness_fab": null,
-    "Wiring_Harness_Install": null,
-    "SHV_splice": null,
-    "CR_install_1st10": null,
-    "CR_install_2nd10": null,
-    "Electric_comment": null,
-    "Elec_comment": null,
-    "Final_Survey": null,
-    "Plane2Plane": null,
-    "Long_inteface_arms": null,
-    "t_brackets": null,
-    "yoke_test": null,
-    "g_bias_boards": null,
-    "final_comment": null,
-    "PD_rail_check": null
-  },
-  {
-    "APAID": "PROTODUNE3",
-    "ID": 5,
-    "CR_Boards_Cryo": "AL/XZ 1/21/18",
-    "Wiring_Harness_fab": "AL/XZ 1/21/18",
-    "Wiring_Harness_Install": "AL/XZ 1/24/18",
-    "SHV_splice": "AL 1/21/18",
-    "CR_install_1st10": "AL 1/25/18 See comments",
-    "CR_install_2nd10": "AL 1/25/18",
-    "Electric_comment": null,
-    "Elec_comment": "14 CR boards installed at PSL, 6 were installed at CERN                      20 G-bias boards were installed 1/25/18                                                   G-level head boards had to bereamed and cleaned to accept the mounting screws for the G-bias boards.  2216 Gray process has been adjusted to eliminate this issue.",
-    "Final_Survey": "1/19/18 MS/QX",
-    "Plane2Plane": "1/19/18 MS",
-    "Long_inteface_arms": "1/22/18 DH",
-    "t_brackets": "1/24/18 MS/PML/JH",
-    "yoke_test": "1/30/18 MS/PML",
-    "g_bias_boards": null,
-    "final_comment": null,
-    "PD_rail_check": "1/24/18 PML"
-  },
-  {
-    "APAID": "US APA 004",
-    "ID": 6,
-    "CR_Boards_Cryo": "NA, not cryo-tested due to schedule",
-    "Wiring_Harness_fab": "3/18/18 Complete",
-    "Wiring_Harness_Install": "3/19/18 AL",
-    "SHV_splice": "3/18/18 Complete",
-    "CR_install_1st10": "3/20/18 AL/XZ",
-    "CR_install_2nd10": "3/20/18 AL/XZ",
-    "Electric_comment": null,
-    "Elec_comment": "G-bias boards and CR boards were not cryo tested.  Schedule shipping constraints and past testing showed no failures resulted after cryo-testing.",
-    "Final_Survey": "3/14/18 MS/QX",
-    "Plane2Plane": "3/14/18 MS/DH",
-    "Long_inteface_arms": "3/15/18 DH",
-    "t_brackets": "3/15/18 MS/JH",
-    "yoke_test": "3/15/18 MS/JH",
-    "g_bias_boards": "3/16/18 JH",
-    "final_comment": null,
-    "PD_rail_check": "3/14/18 JL/PML"
-  },
-  {
-    "APAID": "PROTODUNE1",
-    "ID": 2,
-    "CR_Boards_Cryo": "7/1/17 XZ/AL",
-    "Wiring_Harness_fab": "7/4/17 AL/AA/PML",
-    "Wiring_Harness_Install": "7/5/17 BB/PML/AL",
-    "SHV_splice": "7/5/17 AL/PML",
-    "CR_install_1st10": "7/5/17 AL/PML",
-    "CR_install_2nd10": "7/5/17 AL/PML",
-    "Electric_comment": null,
-    "Elec_comment": null,
-    "Final_Survey": null,
-    "Plane2Plane": null,
-    "Long_inteface_arms": null,
-    "t_brackets": null,
-    "yoke_test": null,
-    "g_bias_boards": null,
-    "final_comment": null,
-    "PD_rail_check": null
-  },
-  {
-    "APAID": "PROTODUNE2",
-    "ID": 3,
-    "CR_Boards_Cryo": "10/24/17 AL",
-    "Wiring_Harness_fab": "10/24-26 KW\r\n10/31-11/1 AL/ XZ",
-    "Wiring_Harness_Install": "11/2/17 AL/XZ/PML",
-    "SHV_splice": "11/2/17 AL",
-    "CR_install_1st10": "11/2/17 AL/PML",
-    "CR_install_2nd10": "11/2/17 AL/PML",
-    "Electric_comment": "11/2/17 20 G Bias boards installed AL",
-    "Elec_comment": "20 G Bias Boards installed 11/2/17 AL                                                      PD Rails- check of 10 openings                                                                    11/3/17 Resulted in 5 openings needing residual epoxy removed to allow PD rails to enter fully                                                11/5/17  5 Slots were cleaned of residual epoxy                                   11/6/17 Re-checked with mock PD module, 10 slots passed                                                                                                                   8 (6 + 2 spares) CR boards shipped to CERN to complete electronics installation",
-    "Final_Survey": "10/31/17",
-    "Plane2Plane": "10/31/17 MS",
-    "Long_inteface_arms": "11/1/17 DH",
-    "t_brackets": "11/2/17",
-    "yoke_test": "11/2/17",
-    "g_bias_boards": null,
-    "final_comment": "2216 Gray Squish out in PD rail locations had to be removed/cleaned to ensure entry into all (10) slots.  Resolution: hold down plates were modified with matching slots.",
-    "PD_rail_check": "11/3-6/17 PML"
-  }
-]
