--- data-documents/prototraveller/sietch_win/tblGeneral.json
@@ -1,47 +1,0 @@
-[
-  {
-    "APAID": "ProcessAPA",
-    "Frame_Serial_Number": "PROTO1",
-    "Head_SN": null,
-    "Center_SN": null,
-    "Side1_SN": null,
-    "Side2_SN": null,
-    "Foot_SN": null
-  },
-  {
-    "APAID": "PROTODUNE1",
-    "Frame_Serial_Number": "001",
-    "Head_SN": "0002",
-    "Center_SN": "0004",
-    "Side1_SN": "0004",
-    "Side2_SN": "0004",
-    "Foot_SN": "0004"
-  },
-  {
-    "APAID": "PROTODUNE2",
-    "Frame_Serial_Number": "002",
-    "Head_SN": "0003",
-    "Center_SN": "0005",
-    "Side1_SN": "0005",
-    "Side2_SN": "0005",
-    "Foot_SN": "0006"
-  },
-  {
-    "APAID": "PROTODUNE3",
-    "Frame_Serial_Number": "003",
-    "Head_SN": "0004",
-    "Center_SN": "0007",
-    "Side1_SN": "0006",
-    "Side2_SN": "0006",
-    "Foot_SN": "0007"
-  },
-  {
-    "APAID": "US APA 004",
-    "Frame_Serial_Number": "004",
-    "Head_SN": "0005",
-    "Center_SN": "0006",
-    "Side1_SN": "0007",
-    "Side2_SN": "0007",
-    "Foot_SN": "0005"
-  }
-]
